--- a/central/cve/cluster/datastore/store/postgres/full_store.go
+++ b/central/cve/cluster/datastore/store/postgres/full_store.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	protoTypes "github.com/gogo/protobuf/types"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/stackrox/rox/central/cve/cluster/datastore/store"
 	"github.com/stackrox/rox/central/cve/converter/v2"
 	"github.com/stackrox/rox/central/metrics"