--- a/migrator/migrations/m_172_to_m_173_network_flows_partition/stores/previous/store.go
+++ b/migrator/migrations/m_172_to_m_173_network_flows_partition/stores/previous/store.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/gogo/protobuf/types"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/stackrox/rox/generated/storage"
 	frozenSchema "github.com/stackrox/rox/migrator/migrations/frozenschema/v73"
 	"github.com/stackrox/rox/pkg/logging"