--- a/migrator/migrations/m_172_to_m_173_network_flows_partition/stores/updated/store.go
+++ b/migrator/migrations/m_172_to_m_173_network_flows_partition/stores/updated/store.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/gogo/protobuf/types"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/postgres"