--- a/migrator/migrations/m_173_to_m_174_group_unique_constraint/stores/previous/postgres_plugin.go
+++ b/migrator/migrations/m_173_to_m_174_group_unique_constraint/stores/previous/postgres_plugin.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/hashicorp/go-multierror"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/pkg/errors"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"