--- a/migrator/migrations/m_179_to_m_180_openshift_policy_exclusions/postgres/store.go
+++ b/migrator/migrations/m_179_to_m_180_openshift_policy_exclusions/postgres/store.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/hashicorp/go-multierror"
-	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v5"
 	"github.com/pkg/errors"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"