## Description

## Description

> pgx follows semantic versioning for the documented public API on stable releases. 
> [**v5 is the latest stable major version**.](https://github.com/jackc/pgx#version-policy)

- https://github.com/go-gorm/postgres/issues/139
- https://github.com/stackrox/stackrox/pull/4676

Important changes in v5:
- `CommandTag` is a type
- [handling arrays – we have code that convert `[]TextArray` but now it's just a `[]string` so only filtering part is needed.](https://github.com/jackc/pgx/blob/v5.4.2/CHANGELOG.md#array-types)
- handling null values in jsonb there was `'null'` but now it's `null` so I added a change to use `{}` with `pgutils.EmptyOrMap`

Since https://github.com/driftprogramming/pgxpoolmock is not maintained over the last year and does not have pgx v5 support I decide to copy mocker rows file to our repo and clean it up from unused functions.

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed

CI