@connorgorman 2023-07-23
Why copy the code that's unmaintained instead of changing the test? Aren't these interfaces that we could write our own mocks for?