@janisz 2023-10-27
Benchmark results
```
benchmark                                                                                                                                                                                                        old ns/op       new ns/op       delta
BenchmarkDatastore_Add_and_Flush/add_1000_events_to_the_writer_and_flush_it-8                                                                                                                                    11335579        15354683        +35.46%
BenchmarkDatastore_Add_and_Flush/add_5000_events_to_the_writer_and_flush_it-8                                                                                                                                    81834037        111430728       +36.17%
BenchmarkDatastore_Add_and_Flush/add_10000_events_to_the_writer_and_flush_it-8                                                                                                                                   167060913       222866560       +33.40%
BenchmarkDatastore_Add_and_Flush/add_20000_events_to_the_writer_and_flush_it-8                                                                                                                                   335972717       458248678       +36.39%
BenchmarkAlertDatabaseOps/searchWithStringHighlighted-8                                                                                                                                                          3475865         3671180         +5.62%
BenchmarkAlertDatabaseOps/searchWithoutHighlighted_(aka_get_IDs_only)-8                                                                                                                                          1279547         1319328         +3.11%
BenchmarkAlertDatabaseOps/searchWithRawListAlert-8                                                                                                                                                               16368310        17039879        +4.10%
BenchmarkAlertDatabaseOps/markResolvedBatch-8                                                                                                                                                                    62763620        72450585        +15.43%
BenchmarkGetMany/1_alerts-8                                                                                                                                                                                      190639          219163          +14.96%
BenchmarkGetMany/2_alerts-8                                                                                                                                                                                      237179          255325          +7.65%
BenchmarkGetMany/4_alerts-8                                                                                                                                                                                      279592          322419          +15.32%
BenchmarkGetMany/8_alerts-8                                                                                                                                                                                      380825          422548          +10.96%
BenchmarkGetMany/16_alerts-8                                                                                                                                                                                     554471          638544          +15.16%
BenchmarkGetMany/32_alerts-8                                                                                                                                                                                     915299          1034500         +13.02%
BenchmarkGetMany/64_alerts-8                                                                                                                                                                                     1678420         1893280         +12.80%
BenchmarkGetMany/128_alerts-8                                                                                                                                                                                    3509091         3873571         +10.39%
BenchmarkGetMany/256_alerts-8                                                                                                                                                                                    6492309         6978606         +7.49%
BenchmarkGetMany/512_alerts-8                                                                                                                                                                                    12291369        13108133        +6.65%
BenchmarkGetMany/1024_alerts-8                                                                                                                                                                                   22732191        25302251        +11.31%
BenchmarkGetMany/2048_alerts-8                                                                                                                                                                                   44885365        49124215        +9.44%
BenchmarkAggregatedResults-8                                                                                                                                                                                     161653844       160989788       -0.41%
BenchmarkSearchAllDeployments/SearchRetrievalList-8                                                                                                                                                              9321001         10966771        +17.66%
BenchmarkSearchAllDeployments/SearchRetrievalFull-8                                                                                                                                                              9016056         10849381        +20.33%
BenchmarkDefaultPolicies-8                                                                                                                                                                                       5935240         6161569         +3.81%
BenchmarkImageResolver/GetImageComponentsInImageScanResolver-8                                                                                                                                                   50688758        52601009        +3.77%
BenchmarkImageResolver/GetImageComponentsWithoutImageScanResolver-8                                                                                                                                              1221931247      1269933524      +3.93%
BenchmarkImageResolver/GetImageComponentsDerivedFieldsWithImageScanResolver-8                                                                                                                                    87350451        89418004        +2.37%
BenchmarkImageResolver/GetImageComponentsDerivedWithoutImageScanResolver-8                                                                                                                                       3094368510      3197060268      +3.32%
BenchmarkImageResolver/GetImageOnly-8                                                                                                                                                                            435997          449065          +3.00%
BenchmarkImageResolver/GetImageWithCounts-8                                                                                                                                                                      15626413        15705896        +0.51%
BenchmarkImageResolver/GetImageScanTimeTopLevel-8                                                                                                                                                                490827          509738          +3.85%
BenchmarkImageResolver/GetImageScanTimeNested-8                                                                                                                                                                  562707          573139          +1.85%
BenchmarkNodeResolver/GetNodeComponentsInNodeScanResolver-8                                                                                                                                                      53771493        55248182        +2.75%
BenchmarkNodeResolver/GetNodeComponentsWithoutNodeScanResolver-8                                                                                                                                                 281134196       295086124       +4.96%
BenchmarkNodeResolver/GetNodeComponentsDerivedFieldsWithNodeScanResolver-8                                                                                                                                       89534816        91711970        +2.43%
BenchmarkNodeResolver/GetNodeComponentsDerivedWithoutNodeScanResolver-8                                                                                                                                          1954803051      2099329933      +7.39%
BenchmarkNodeResolver/GetNodeOnly-8                                                                                                                                                                              466761          476666          +2.12%
BenchmarkNodeResolver/GetNodeWithCounts-8                                                                                                                                                                        10547056        23316144        +121.07%
BenchmarkImageGetMany/GetImagesBatch-8                                                                                                                                                                           102263803       109245059       +6.83%
BenchmarkImageGetMany/GetManyImageMetadata-8                                                                                                                                                                     1453211         1552333         +6.82%
BenchmarkInitFromStore/New-8                                                                                                                                                                                     260400844       292017617       +12.14%
BenchmarkNetEntityCreates/createNetworkEntities-8                                                                                                                                                                4.75            4.78            +0.67%
BenchmarkNetEntityCreateBatch/createNetworkEntitiesBatch-8                                                                                                                                                       4.58            23251           +507896.50%
BenchmarkNetEntityUpdates/updateNetworkEntities-8                                                                                                                                                                2.85            2.96            +3.82%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_1_deployments;_1_policies-8                                                                                                                                 5646            5743            +1.72%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_1_deployments;_10_policies-8                                                                                                                                12124           12442           +2.62%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_1_deployments;_100_policies-8                                                                                                                               77055           77557           +0.65%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_10_deployments;_1_policies-8                                                                                                                                16631           17131           +3.01%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_10_deployments;_10_policies-8                                                                                                                               33905           33855           -0.15%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_10_deployments;_100_policies-8                                                                                                                              203731          205162          +0.70%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_100_deployments;_1_policies-8                                                                                                                               130112          134809          +3.61%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_100_deployments;_10_policies-8                                                                                                                              250812          255026          +1.68%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_100_deployments;_100_policies-8                                                                                                                             1447300         1453385         +0.42%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_1000_deployments;_1_policies-8                                                                                                                              1472829         1496591         +1.61%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_1000_deployments;_10_policies-8                                                                                                                             2663612         2710558         +1.76%
BenchmarkEvaluateCluster/_BenchmarkEvaluateCluster_-_1000_deployments;_100_policies-8                                                                                                                            14741152        14714595        -0.18%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_200_ingress_allowed_per_policy_200_egress_allowed_per_policy-8                                                                      123927895       126904223       +2.40%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_200_ingress_allowed_per_policy_1000_egress_allowed_per_policy-8                                                                     352366369       5875395415      +1567.41%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_200_ingress_allowed_per_policy_2000_egress_allowed_per_policy-8                                                                     5943010363      6048590197      +1.78%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_1000_ingress_allowed_per_policy_200_egress_allowed_per_policy-8                                                                     5349449365      385703312       -92.79%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_1000_ingress_allowed_per_policy_1000_egress_allowed_per_policy-8                                                                    7746303919      5906863660      -23.75%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_1000_ingress_allowed_per_policy_2000_egress_allowed_per_policy-8                                                                    11619499227     12468024236     +7.30%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_2000_ingress_allowed_per_policy_200_egress_allowed_per_policy-8                                                                     7855350256      8239130123      +4.89%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_2000_ingress_allowed_per_policy_1000_egress_allowed_per_policy-8                                                                    8706379878      8632572057      -0.85%
BenchmarkDensePolicies/_BenchmarkDensePolicies_-_2000_deployments_1_policies_2000_ingress_allowed_per_policy_2000_egress_allowed_per_policy-8                                                                    12052575937     12798788916     +6.19%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_1_policies_each_applied_to_1_nodes-8                                                                                                       31859966        33682849        +5.72%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_1_policies_each_applied_to_10_nodes-8                                                                                                      31496210        33751554        +7.16%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_1_policies_each_applied_to_100_nodes-8                                                                                                     32051465        33529666        +4.61%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_10_policies_each_applied_to_1_nodes-8                                                                                                      168840294       172083732       +1.92%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_10_policies_each_applied_to_10_nodes-8                                                                                                     170410384       187238559       +9.88%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_10_policies_each_applied_to_100_nodes-8                                                                                                    177099805       185351952       +4.66%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_100_policies_each_applied_to_1_nodes-8                                                                                                     1516839155      1566172655      +3.25%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_100_policies_each_applied_to_10_nodes-8                                                                                                    1554830224      1570951010      +1.04%
BenchmarkSparsePolicies/_BenchmarkSparsePolicies_-_10000_deployments;_100_policies_each_applied_to_100_nodes-8                                                                                                   1834543487      1940255677      +5.76%
BenchmarkGetManyNodes/GetNodesBatch-8                                                                                                                                                                            94202716        97658165        +3.67%
BenchmarkGetManyNodes/GetManyNodeMetadata-8                                                                                                                                                                      982009          1311310         +33.53%
BenchmarkNodes/upsertNodeWithOldScan-8                                                                                                                                                                           1467426         1511937         +3.03%
BenchmarkNodes/upsertNodeWithNewScan-8                                                                                                                                                                           1474778         1538221         +4.30%
BenchmarkNodes/searchAll-8                                                                                                                                                                                       0.15            0.15            -1.62%
…